% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census.R
\name{get_census}
\alias{get_census}
\title{Get census}
\usage{
get_census(lscape, keep_output, api_key = Sys.getenv("CENSUS_API_KEY"))
}
\arguments{
\item{api_key}{}
}
\description{
This function takes an input lakescape, a user-defined land area around a 
lake, and returns the total population for that area.
}
\examples{
get_census()
}
