% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_impervious.R
\name{get_impervious}
\alias{get_impervious}
\title{Get impervious}
\usage{
get_impervious(lake, lakescape, geo_data)
}
\arguments{
\item{geo_data}{}
}
\description{
Get impervious
}
\examples{
get_impervious()
}
