% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lakescape.R
\name{get_lakescape}
\alias{get_lakescape}
\title{Get lakescape}
\usage{
get_lakescape(lake, lakescape, geo_data)
}
\arguments{
\item{geo_data}{}
}
\description{
Get lakescape
}
\examples{
get_lakescape()
}

