% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_lakescape.R
\name{make_lakescape}
\alias{make_lakescape}
\title{Make lakescape}
\usage{
make_lakescape(lake, lakescape, geo_data)
}
\arguments{
\item{geo_data}{}
}
\description{
Make lakescape
}
\examples{
make_lakescape()
}

