% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lakescape.R
\name{get_lakescape}
\alias{get_lakescape}
\title{Get lakescape}
\usage{
get_lakescape(lake, lakescape, geo_data)
}
\arguments{
\item{geo_data}{}
}
\description{
The \code{get_lakescape} function uses a lake and a user defined landscape 
(i.e. the lakescape) to summarize various sources of ancillary geospatial 
data.
}
\examples{
get_lakescape()
}

